+++ pajbot/apiwrappers/twitch/helix.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional, List
+from typing import Dict, Optional, List, Tuple
 
 import logging
 import time
@@ -9,8 +9,14 @@ import math
 from requests import HTTPError
 
 from pajbot import utils
-from pajbot.apiwrappers.response_cache import DateTimeSerializer, ClassInstanceSerializer, ListSerializer
+from pajbot.apiwrappers.response_cache import (
+    DateTimeSerializer,
+    ClassInstanceSerializer,
+    ListSerializer,
+    TwitchChannelEmotesSerializer,
+)
 from pajbot.apiwrappers.twitch.base import BaseTwitchAPI
+from pajbot.models.emote import Emote
 from pajbot.models.user import UserBasics, UserChannelInformation, UserStream
 from pajbot.utils import iterate_in_chunks
 
@@ -537,3 +543,66 @@ class TwitchHelixAPI(BaseTwitchAPI):
             raise e
 
         return response.status_code == 204
+
+    def fetch_global_emotes(self) -> List[Emote]:
+        # circular import prevention
+        from pajbot.managers.emote import EmoteManager
+
+        resp = self.get("/chat/emotes/global")
+
+        return [EmoteManager.twitch_emote(str(emote["id"]), emote["name"]) for emote in resp["data"]]
+
+    def get_global_emotes(self, force_fetch=False) -> List[Emote]:
+        return self.cache.cache_fetch_fn(
+            redis_key="api:twitch:helix:global-emotes",
+            fetch_fn=lambda: self.fetch_global_emotes(),
+            serializer=ListSerializer(Emote),
+            expiry=60 * 60,
+            force_fetch=force_fetch,
+        )
+
+    def fetch_channel_emotes(self, channel_id: str, channel_name: str) -> Tuple[List[Emote], List[Emote], List[Emote]]:
+        """Returns a tuple of three lists of emotes, each one corresponding to tier 1, tier 2 and tier 3 respectively.
+        Tier 2 and Tier 3 ONLY contain the respective extra emotes added to that tier."""
+        # circular import prevention
+        from pajbot.managers.emote import EmoteManager
+
+        resp = self.get("/chat/emotes", {"broadcaster_id": channel_id})
+
+        emotes = resp["data"]
+        if len(emotes) <= 0:
+            log.warning(f"No subscription emotes found for channel {channel_name}")
+            return [], [], []
+
+        ret_data: Tuple[List[Emote], List[Emote], List[Emote]] = ([], [], [])
+        for emote in emotes:
+            if emote["emote_type"] == "subscriptions":
+                tier = 0
+                if emote["tier"] == "1000":  # tier 1 emotes
+                    tier = 1
+                elif emote["tier"] == "2000":  # tier 2 emotes
+                    tier = 2
+                elif emote["tier"] == "3000":  # tier 3 emotes
+                    tier = 3
+                else:
+                    log.warning(f"Unknown channel emote tier fetched: '{emote}'")
+                    continue
+
+                ret_data[tier - 1].append(EmoteManager.twitch_emote(str(emote["id"]), emote["name"]))
+            elif emote["emote_type"] == "bitstier":
+                # TODO: Figure out where bit emotes fit into pajbot
+                pass
+            else:
+                log.warning(f"Unknown channel emote type fetched: '{emote}'")
+        return ret_data
+
+    def get_channel_emotes(
+        self, channel_id: str, channel_name: str, force_fetch=False
+    ) -> Tuple[List[Emote], List[Emote], List[Emote]]:
+        return self.cache.cache_fetch_fn(
+            redis_key=f"api:twitch:helix:channel-emotes:{channel_name}",
+            fetch_fn=lambda: self.fetch_channel_emotes(channel_id, channel_name),
+            serializer=TwitchChannelEmotesSerializer(),
+            expiry=60 * 60,
+            force_fetch=force_fetch,
+        )